implemented refresh token

+++ sid-address-verification/src/main/java/tech/sourceid/sid_address_verification/data/responses/GetOrganisationConfigResponse.kt
@@ -8,7 +8,7 @@ data class GetOrganisationConfigResponse(
 )
 
 data class OrganisationConfigData(
-    val distanceTolerance: Int,
+    val distanceTolerance: Double,
     val geotaggingPollingInterval: Double,
     val geotaggingSessionTimeout: Int
 )
